Use reqwest in place of hyper & destructure structs
This is a larger rewrite removing codegen dependencies and the direct
usage of hyper. Instead we use reqwest to simplify the http request.
We also destructure the structs that we deserialize our JSON against and
do pattern matching on the struct fields. This way we avoid debug
printing, and as a result avoid printing: Some(Package number): Package
Number..

The output is cleaner, however the implementation is a bit ugly.